Introduce support for NextHop type handling (#43)
* Introduce support for NextHop type handling

Command line client can generate different types of
NextHops now. (STD, LB, NAT)

Signed-off-by: Guvenc Gulce <guevenc.guelce@sap.com>

* LB Routes should be distributed back to the originator

As a LoadBalancer can exist together with a NetworkInterface
on the same Hypervisor and share the same metalbond client,
the LB announcement of the NetworkInterface should be reflected
back to originator too. Otherwise a possible LoadBalancer
on the same Hypervisor node doesnt have a chance to receive
the LB announcement of that NetworkInterface.

Signed-off-by: Guvenc Gulce <guevenc.guelce@sap.com>

Signed-off-by: Guvenc Gulce <guevenc.guelce@sap.com>